# specs/001-processo-seletivo-editais/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Processo Seletivo e Editais API
  version: 1.0.0-draft
  description: Contrato de planejamento; não representa implementação existente.
servers:
  - url: /api/v1
...
paths:
...
  /public/publicacoes/{publicacaoId}:
    get:
      tags: [Consulta pública]
      summary: Consultar metadados e conteúdo de uma Publicação imutável
      operationId: consultarPublicacao
      parameters: [{ $ref: '#/components/parameters/PublicacaoId' }]
      responses:
        '200': { description: Publicação histórica, content: { application/json: { schema: { $ref: '#/components/schemas/PublicacaoDetalheResponse' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A autenticação institucional será definida no planejamento de integração; cada operação ainda exige permissão contextual no backend.
  parameters:
...
    PublicacaoId: { name: publicacaoId, in: path, required: true, schema: { $ref: '#/components/schemas/Id' } }
...
  responses:
...
    NotFound: { description: Recurso inexistente ou não revelado ao ator, content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
...
  schemas:
    Id: { type: string, format: uuid }
...
    SignatorySnapshot:
      type: object
      additionalProperties: false
      required: [authorityId, name, role]
      properties:
        authorityId: { $ref: '#/components/schemas/Id' }
        name: { type: string, minLength: 1 }
        role: { type: string, minLength: 1 }
...
    PublicacaoResponse:
      type: object
      required: [id, editalId, publicationOrder, publishedAt, effectiveAt, contentHash, documentHash]
      properties:
        kind: { type: string, enum: [PUBLICACAO], description: Discrimina o item dentro de HistoricoPage }
        id: { $ref: '#/components/schemas/Id' }
        editalId: { $ref: '#/components/schemas/Id' }
        publicationOrder: { type: integer, format: int64 }
        publishedAt: { type: string, format: date-time }
        effectiveAt: { type: string, format: date-time }
        contentHash: { type: string }
        documentHash: { type: string }
    PublicacaoDetalheResponse:
      allOf:
        - $ref: '#/components/schemas/PublicacaoResponse'
        - type: object
          required: [sourceType, sourceId, signatory, content, documentUrl]
          properties:
            sourceType: { type: string, enum: [EDITAL, RETIFICACAO] }
            sourceId: { $ref: '#/components/schemas/Id' }
            signatory: { $ref: '#/components/schemas/SignatorySnapshot' }
            content: { type: object, additionalProperties: true }
            documentUrl: { type: string, format: uri-reference }
...
    Problem:
      type: object
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        instance: { type: string }
        correlationId: { type: string }
        violations: { type: array, items: { type: object } }